Major changes (MongoDB, Login, Signup) check disc.
Added the mongoDB connection, login and signup APIs with bcrypt.
The remaining parts to do would be tying in the API call to
OpenWeatherMap. Idea is the user will enter either a city or ZIP code
(might just skip the city and just do ZIP code) and when they click
Find, it'll do a drop down of the location they want to add. If
there are already 3, the Find button will be disabled. If less, it will
be added to the cards. Then need to implement the delete location.