Implemented part of the OpenWeatherAPI calls.
Was more difficult than expect. OWM API doesn't have a API call that
solely depends on zip code. It needs both long and lat along with zip.
So the idea now is to do a API call on acquiring long and lat for each
zip, then do a call after getting long and lat.